// frontend/lib/api/generated/model/salesActivityRequestCreateRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Orbis API
 * Orbis API
 * OpenAPI spec version: 1.0.0
 */
import type { SalesActivityRequestCreateRequestActivityPurpose } from './salesActivityRequestCreateRequestActivityPurpose';

export interface SalesActivityRequestCreateRequest {
  targetUserId?: string;
  activityPurpose?: SalesActivityRequestCreateRequestActivityPurpose;
  activityDateTime?: string;
  requestContent?: string;
}