// frontend/lib/api/generated/model/salesActivityRequestResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Orbis API
 * Orbis API
 * OpenAPI spec version: 1.0.0
 */
import type { SalesActivityRequestResponseActivityPurpose } from './salesActivityRequestResponseActivityPurpose';

export interface SalesActivityRequestResponse {
  id?: number;
  salesActivityId?: number;
  targetUserId?: string;
  activityPurpose?: SalesActivityRequestResponseActivityPurpose;
  activityDateTime?: string;
  requestContent?: string;
}