// frontend/lib/api/generated/model/salesActivityRequestResponseActivityPurpose.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Orbis API
 * Orbis API
 * OpenAPI spec version: 1.0.0
 */

export type SalesActivityRequestResponseActivityPurpose = typeof SalesActivityRequestResponseActivityPurpose[keyof typeof SalesActivityRequestResponseActivityPurpose];


export const SalesActivityRequestResponseActivityPurpose = {
  CONSULTING: 'CONSULTING',
  PRODUCT_INTRODUCTION: 'PRODUCT_INTRODUCTION',
  DEMO: 'DEMO',
  POC: 'POC',
  BMT: 'BMT',
  DOCUMENT_DELIVERY: 'DOCUMENT_DELIVERY',
  RFP_ANALYSIS: 'RFP_ANALYSIS',
  PROPOSAL_WRITING: 'PROPOSAL_WRITING',
  SI_PROPOSAL_WRITING: 'SI_PROPOSAL_WRITING',
  ETC: 'ETC',
} as const;